Версия 3.0.23 [!] Появилась опция $_CONFIG['CACHE_MODE'], позволяющая подключать разные классы, отвечающие за кэширование. По умолчанию $_CONFIG['CACHE_MODE']="base", соответственно, подключается класс HolyCache_base. [!] Добавлена вариант для опции $_CONFIG['CACHE_MODE']="files" - хранение кэша в файлах. [!] Решена проблема кэширование комплексных компонентов - таким компонентам необходимо передавать параметр inner_cahce, а не cache. [*] Поправлены ошибки в кэшировании

+++ admin/clear_cache.php
@@ -5,7 +5,8 @@ if (isset($_POST['form_go'])) {
 
     if (isset($_POST['clear_base'])) {
         $ok[] = "Кэш базы очищен";
-        $SQL->Query('TRUNCATE TABLE `system_cache`');
+        $_tmp_holy_cache=new HolyCacheOut();
+        $_tmp_holy_cache->ClearFull();
     };
     if (isset($_POST['clear_resize'])) {
         $ok[] = "Кэш картинок очищен";
@@ -44,25 +45,29 @@ if (isset($_POST['form_go'])) {
                     $elements->GetList();
                     //получаем непосредственно картинки
                     while ($tmp = $elements->GetNext()) {
-                        $tmp = explode(";",$tmp[$field['name']]);
+                        $tmp = explode(";", $tmp[$field['name']]);
                         foreach ($tmp as $_img)
                             if ($_img) {
-                                $pic[] = FOLDER_ROOT.$_img;
+                                $pic[] = FOLDER_ROOT . $_img;
                             }
                     }
                 };
         };
         //ищем реальные файлы
-        $real_pics=FileFindRecursive(FOLDER_IMAGE,Array("jpeg","jpg","gif","png"));
-        
-        //сравниваем списки реальных файлов и сохраненных в базе
-        $need_to_delete=array_diff($real_pics,$pic);
-        
-        //удаляем
-        foreach ($need_to_delete as $_delete){
-            unlink($_delete);
-        }
-        
+        $real_pics = FileFindRecursive(FOLDER_IMAGE, Array("jpeg", "jpg", "gif", "png"));
+        if (is_array($pic))
+            if (is_array($real_pics)) {
+                //сравниваем списки реальных файлов и сохраненных в базе
+                $need_to_delete = array_diff($real_pics, $pic);
+
+                //удаляем
+                if ($need_to_delete)
+                    if (is_array($need_to_delete)) {
+                        foreach ($need_to_delete as $_delete) {
+                            unlink($_delete);
+                        }
+                    };
+            };
         $ok[] = "Картинки-сиротки удалены";
     };
 };
@@ -83,21 +88,21 @@ if (isset($ok)) {
 <form method=post>
     <input type=hidden name="form_go" value="512">
     Выберите, что очищать:<BR><BR>
-<?
-global $_CONFIG;
-if ($_CONFIG['CACHE_SYSTEM']) {
-    ?>
+    <?
+    global $_CONFIG;
+    if ($_CONFIG['CACHE_SYSTEM']) {
+        ?>
         <label class="checkbox"><input checked type=checkbox name="clear_base" value="1"> &nbsp; Очистить кэш базы</label>
-     
+
     <? }; ?>
     <label class="checkbox"><input checked type=checkbox name="clear_resize" value="1"> &nbsp; Удалить кэш картинок</label>
-   
+
     <label class="checkbox"><input checked type=checkbox name="clear_tmp" value="1"> &nbsp; Удалить временные файлы</label>
 
-   <label class="checkbox"> <input checked type=checkbox name="clear_orphan" value="1"> &nbsp; Удалить картинки-сиротки</label>
-<? /* <BR>
-  <input type=checkbox name="clear_orphan" value="1"> &nbsp; Удалить несуществующие картинки (долгий процесс, но позволяет сэкономить место)
- */ ?>
+    <label class="checkbox"> <input checked type=checkbox name="clear_orphan" value="1"> &nbsp; Удалить картинки-сиротки</label>
+    <? /* <BR>
+      <input type=checkbox name="clear_orphan" value="1"> &nbsp; Удалить несуществующие картинки (долгий процесс, но позволяет сэкономить место)
+     */ ?>
 
     </br><br><input name=submit type=submit value="Очистить кэш" style="width:40%;HEIGHT:28px;" class="btn btn-success">
 </form>
